Refactoring ECS code, restructure files
